requirements.txt
sanic
reportlab
requests
aiohttp
jinja2
Pillow
